Do not verify the signature while use adb to install EPK
Verify the DID signature when the EPK use the new method to sign.
Compartiable with the lagacy signature for now.

Will remove the passphrase related code for opening a DID store when
the DID Java SDK supports open DID store without passphrase.

TODO: Resolve the DID URL from the chain.
      Pack the DID doc into EPK and self prove, self verify the DID signature.